+++ src/commands/init.js
@@ -21,7 +21,7 @@ import {
   ensureUnitTestPolicy,
   UNIT_TEST_POLICY_FILE,
 } from '../policies/managed-policies.js';
-import { ensureProjectConfig } from '../config-management.js';
+import { ensureProjectConfig } from '../orchestration/setup/config-management.js';
 import { detectProjectBuildSetup } from '../gates/quality/build-setup.js';
 import { findRepositoryRoot } from '../git.js';
 import { installHooks } from '../orchestration/setup/hook-installer.js';